util: split out signal-util.[ch] from util.[ch]
No functional changes.
