+++ src/core/dbus-kill.c
@@ -19,9 +19,11 @@
   along with systemd; If not, see <http://www.gnu.org/licenses/>.
 ***/
 
+#include "signal-util.h"
+#include "bus-util.h"
+
 #include "kill.h"
 #include "dbus-kill.h"
-#include "bus-util.h"
 
 static BUS_DEFINE_PROPERTY_GET_ENUM(property_get_kill_mode, kill_mode, KillMode);
