+++ src/journal-remote/journal-upload.c
@@ -33,8 +33,9 @@
 #include "mkdir.h"
 #include "conf-parser.h"
 #include "sigbus.h"
-#include "journal-upload.h"
 #include "formats-util.h"
+#include "signal-util.h"
+#include "journal-upload.h"
 
 #define PRIV_KEY_FILE CERTIFICATE_ROOT "/private/journal-upload.pem"
 #define CERT_FILE     CERTIFICATE_ROOT "/certs/journal-upload.pem"
