+++ src/libsystemd/sd-event/test-event.c
@@ -23,6 +23,7 @@
 #include "log.h"
 #include "util.h"
 #include "macro.h"
+#include "signal-util.h"
 
 static int prepare_handler(sd_event_source *s, void *userdata) {
         log_info("preparing %c", PTR_TO_INT(userdata));
