+++ src/login/loginctl.c
@@ -43,6 +43,7 @@
 #include "verbs.h"
 #include "process-util.h"
 #include "terminal-util.h"
+#include "signal-util.h"
 
 static char **arg_property = NULL;
 static bool arg_all = false;
