+++ src/machine/machinectl.c
@@ -54,6 +54,7 @@
 #include "import-util.h"
 #include "process-util.h"
 #include "terminal-util.h"
+#include "signal-util.h"
 
 static char **arg_property = NULL;
 static bool arg_all = false;
