+++ src/run/run.c
@@ -35,6 +35,7 @@
 #include "calendarspec.h"
 #include "ptyfwd.h"
 #include "formats-util.h"
+#include "signal-util.h"
 
 static bool arg_scope = false;
 static bool arg_remain_after_exit = false;
