+++ src/tty-ask-password-agent/tty-ask-password-agent.c
@@ -44,6 +44,7 @@
 #include "def.h"
 #include "process-util.h"
 #include "terminal-util.h"
+#include "signal-util.h"
 
 static enum {
         ACTION_LIST,
